Avoid skipping publish in workflow_dispatch
Publish dev version only for `latestMillDevVersion`.
We were not publishing dev versions of mill-mima for every commit
So I'm removing the stable versions of mill from the ones published
with the manual trigger of `workflow_dispatch`.
This way there are not so many versions to choose from in Maven central.